Initial v6 support

+++ rbl.h
@@ -6,4 +6,4 @@ typedef struct rbl_negcache {
 void rbl_init(void);
 void expire_rbl_negcache(int, short, rbl_negcache_t *);
 void get_rbl_answer(int, char, int, int, struct in_addr *, uint32_t *);
-void make_rbl_query(uint32_t addr);
+void make_rbl_query(uint128_t addr);
